+++ feature/report/src/main/java/com/hankki/feature/report/main/ReportRoute.kt
@@ -1,4 +1,4 @@
-package com.hankki.feature.report
+package com.hankki.feature.report.main
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
@@ -53,7 +53,7 @@ import com.hankki.core.designsystem.theme.HankkijogboTheme
 import com.hankki.core.designsystem.theme.Red
 import com.hankki.core.designsystem.theme.White
 import com.hankki.domain.report.entity.CategoryEntity
-import com.hankki.feature.report.model.MenuModel
+import com.hankki.feature.report.main.model.MenuModel
 import kotlinx.collections.immutable.PersistentList
 import kotlinx.collections.immutable.persistentListOf
 import kotlinx.coroutines.delay
