+++ feature/report/src/main/java/com/hankki/feature/report/main/ReportViewModel.kt
@@ -1,8 +1,8 @@
-package com.hankki.feature.report
+package com.hankki.feature.report.main
 
 import androidx.lifecycle.ViewModel
 import com.hankki.domain.report.entity.CategoryEntity
-import com.hankki.feature.report.model.MenuModel
+import com.hankki.feature.report.main.model.MenuModel
 import kotlinx.collections.immutable.persistentListOf
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
