+++ feature/report/src/main/java/com/hankki/feature/report/navigation/ReportNavigation.kt
@@ -4,7 +4,7 @@ import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
 import com.hankki.core.navigation.MainTabRoute
-import com.hankki.feature.report.ReportRoute
+import com.hankki.feature.report.main.ReportRoute
 import kotlinx.serialization.Serializable
 
 fun NavController.navigateReport() {
